import sort & move models to a directory
