+++ netbox_access_lists/choices.py
@@ -8,6 +8,7 @@ __all__ = (
     'ACLActionChoices',
     'ACLRuleActionChoices',
     'ACLTypeChoices',
+    'ACLProtocolChoices',
 )
