+++ netbox_access_lists/graphql.py
@@ -3,8 +3,9 @@ Define the object types and queries availble via the graphql api.
 """
 
 from graphene import ObjectType
-from netbox.graphql.types import NetBoxObjectType
 from netbox.graphql.fields import ObjectField, ObjectListField
+from netbox.graphql.types import NetBoxObjectType
+
 from . import filtersets, models
 
 __all__ = (
