+++ netbox_access_lists/tables.py
@@ -3,8 +3,8 @@ Define the object lists / table view for each of the plugin models.
 """
 
 import django_tables2 as tables
+from netbox.tables import ChoiceFieldColumn, NetBoxTable, columns
 
-from netbox.tables import NetBoxTable, columns, ChoiceFieldColumn
 from .models import AccessList, ACLExtendedRule, ACLStandardRule
 
 __all__ = (
