+++ netbox_access_lists/urls.py
@@ -3,10 +3,9 @@ Map Views to URLs.
 """
 
 from django.urls import path
-
 from netbox.views.generic import ObjectChangeLogView
-from . import models, views
 
+from . import models, views
 
 urlpatterns = (
