+++ netbox_access_lists/views.py
@@ -4,8 +4,8 @@ Specifically, all the various interactions with a client.
 """
 
 from django.db.models import Count
-
 from netbox.views import generic
+
 from . import filtersets, forms, models, tables
 
 __all__ = (
