+++ setup.py
@@ -2,7 +2,7 @@
 #import os.path
 #
 from setuptools import find_packages, setup
-#
+
 #
 #with open("README.md", "r") as fh:
 #    long_description = fh.read()
